requirements.txt
cdsapi
xarray
netCDF4
numpy 
python-dotenv 
psycopg2
timedelta
geopandas
mercantile
geoalchemy2
apache-airflow
apache-airflow-providers-slack
pandas==2.2.*
SQLAlchemy==2.0.*
